# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Shopping Cart",
        "version": "0.0.1"
    },
    "servers": [
        {
            "url": "http://localhost/api"
        }
    ],
    "paths": {
        "/api/carts": {
            "post": {
                "tags": [
                    "Carts"
                ],
                "summary": "Create a new cart",
                "operationId": "b5e65bb90b5ff6c7f38b3417c6570289",
                "requestBody": {
                    "description": "Cart data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "total_price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 0
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Cart created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "cart created successfully"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "500": {
                        "description": "Failed to create cart"
                    }
                }
            }
        },
        "/api/cart": {
            "put": {
                "tags": [
                    "Carts"
                ],
                "summary": "Update a product",
                "operationId": "ddd8c5fe65a398fdb2cbd0dff83a4ea5",
                "requestBody": {
                    "description": "Cart data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "total_price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 0
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Cart updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Cart updated successfully"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "500": {
                        "description": "Failed to update cart"
                    }
                }
            }
        },
        "/api/carts/{id}": {
            "delete": {
                "tags": [
                    "Carts"
                ],
                "summary": "Delete a cart",
                "operationId": "a23b2ec2c218b0cd9ab0114a2f818f17",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the product to delete",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Cart deleted",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Deleted success"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to delete cart"
                    }
                }
            }
        },
        "/api/cart_item": {
            "put": {
                "tags": [
                    "Cart_item"
                ],
                "summary": "Create a new cart_item",
                "operationId": "6609c7258e1c0e0c1f9b5dc026ddbad0",
                "requestBody": {
                    "description": "Cart_item data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "cart_item_id": {
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "cart_id": {
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "product_id": {
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "quantity": {
                                        "type": "integer",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Product created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Cart item updated successfully"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "500": {
                        "description": "Failed to update cart item"
                    }
                }
            },
            "post": {
                "tags": [
                    "Cart_item"
                ],
                "summary": "Create a new cart_item",
                "operationId": "c0822010920321fa1e2986e2fa384e79",
                "requestBody": {
                    "description": "Cart_item data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "cart_id": {
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "product_id": {
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "quantity": {
                                        "type": "integer",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Product created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Cart item created successfully"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "500": {
                        "description": "Failed to create cart item"
                    }
                }
            }
        },
        "/api/cart_item/{id}": {
            "delete": {
                "tags": [
                    "Cart_item"
                ],
                "summary": "Delete a cart item",
                "operationId": "0d0a03ac4d7e017d9a162e9dd1584220",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the cart item to delete",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Cart item deleted",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Deleted success"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to delete cart item"
                    }
                }
            }
        },
...
    }
}